Separate model for single and multiple documents
Either a manifest has one document containing the title, nonsensitive title and description,
_or_ the manifest itself specifies title, nonsensitive title and
description, and has one or more documents with just a document title.

Using some tricks with JAXB and private interfaces to "inject"
code to be able to treat a strictly single document manifest (p.t. direct-job manifest)
as a multiple documents, even though the generated JAXB-classes does not
support this yet. (Reverting the cardinality of the direct-document from
"unbounded" to 1)

The method XMLManifest.getDocument() is appropriately marked as
@Deprecated, and can be replaced with the unified API for multiple
documents.