Use old JAXB artifacts
To ensure the JAXB facilities (generated code and marshallers) can be
living alongside any jakarta.xml.bind namespaced JAXB in newer Java
applications.

This ensures that we are using JAXB API and runtime artifacts which have
other Maven coordinates than the newer Jakarta-branded artifacts, which
started out with the old javax.xml.bind package, but in later versions
changed this to jakarta.xml.bind. If dependees use newer Jakarta JAXB
with the proper package, this will break the use of this library if it
is depending on the Jakarta-artifacts with the javax.xml.bind package.

It's a mess.