Modularizing and upgrading JAXB XJC compilation
Upgrades maven-jaxb2-plugin from 0.14.0 to 0.15.3. This upgrade broke
the current rig for piecing together the XSDs, and

Now, while gazing at the ton of increase in glorious XML, you ask
yourself: "why is all this worth the upgrade from 0.14 to 0.15?".

Well, JAXB gets messy and fragile once you have a non-trivial case, as
this is: mixed compilation of third party XSDs together with your own.
It would seem that the successful operation of the previous setup may
have been a bit accidental. The web of imports, includes, namespacing,
catalogs, and what not which is used to express this API schema was not
resolved as expected by the updated version of the plugin (JAXB, xjc, etc)
and produced duplicates which caused compilation errors.

This change is an effort to try to stay a bit ahead of the game, to
maybe also be able to upgrade to Jakarta JAXB down the road. This is
so-called "modularized XSD compilation" where the third party XSDs we
use are compiled separately, and the output of each compilation sted are
included as "episodes" in further compilation. This seems to resolve
that schemas which were imported by multiple other schemas are now only
compiled once.